Refactor the exceptions
