+++ api/src/main/java/com/michelin/ns4kafka/controllers/AccessControlListController.java
@@ -5,7 +5,7 @@ import com.michelin.ns4kafka.models.Namespace;
 import com.michelin.ns4kafka.models.ObjectMeta;
 import com.michelin.ns4kafka.repositories.AccessControlEntryRepository;
 import com.michelin.ns4kafka.repositories.NamespaceRepository;
-import com.michelin.ns4kafka.validation.ResourceValidationException;
+import com.michelin.ns4kafka.exception.ResourceValidationException;
 import io.micronaut.http.HttpStatus;
 import io.micronaut.http.annotation.*;
 import io.swagger.v3.oas.annotations.Operation;
