+++ api/src/main/java/com/michelin/ns4kafka/controllers/TopicController.java
@@ -13,7 +13,7 @@ import com.michelin.ns4kafka.repositories.AccessControlEntryRepository;
 import com.michelin.ns4kafka.repositories.NamespaceRepository;
 import com.michelin.ns4kafka.repositories.TopicRepository;
 import com.michelin.ns4kafka.services.KafkaAsyncExecutor;
-import com.michelin.ns4kafka.validation.ResourceValidationException;
+import com.michelin.ns4kafka.exception.ResourceValidationException;
 
 import io.micronaut.context.ApplicationContext;
 import io.micronaut.http.HttpResponse;
