+++ api/src/main/java/com/michelin/ns4kafka/exception/ResourceValidationException.java
@@ -1,0 +1,15 @@
+package com.michelin.ns4kafka.exception;
+
+import lombok.Getter;
+
+import java.util.List;
+
+public class ResourceValidationException extends RuntimeException {
+    private static final long serialVersionUID = 32400191899153204L;
+    @Getter
+    List<String> validationErrors;
+
+    public ResourceValidationException(List<String> validationsErrors) {
+        this.validationErrors = validationsErrors;
+    }
+}
