+++ api/src/main/java/com/michelin/ns4kafka/exceptionHandler/ResourceValidationExceptionHandler.java
@@ -1,22 +1,23 @@
 package com.michelin.ns4kafka.exceptionHandler;
 
+import javax.inject.Singleton;
+
+import com.michelin.ns4kafka.exception.ResourceValidationException;
+
 import io.micronaut.context.annotation.Requires;
 import io.micronaut.http.HttpRequest;
 import io.micronaut.http.HttpResponse;
 import io.micronaut.http.annotation.Produces;
 import io.micronaut.http.server.exceptions.ExceptionHandler;
-import javax.inject.Singleton;
-
-import com.michelin.ns4kafka.validation.ResourceValidationException;
 
 @Produces
 @Singleton
 @Requires(classes = { ResourceValidationException.class, ExceptionHandler.class })
 public class ResourceValidationExceptionHandler
-        implements ExceptionHandler<ResourceValidationException, HttpResponse<T>> {
+        implements ExceptionHandler<ResourceValidationException, HttpResponse<ResourceCreationError>> {
 
     @Override
-    public HttpResponse<ResourceCreationError> handle(HttpRequest<T> request, ResourceValidationException exception) {
+    public HttpResponse<ResourceCreationError> handle(HttpRequest request, ResourceValidationException exception) {
         return HttpResponse.badRequest()
                 .body(new ResourceCreationError("Message validation failed", exception.getValidationErrors()));
     }
